end of chapter 3. the bvh created a huge speed up for the runtime of the raytracer as a whole. a particular scenario took about 30 minutes but now takes > 5 minutes with the bvh, and all on one core.
Coming up with algorithms for how to create the bvh would be very cool. The creation process is fast, but how the elements are organized might result in better distributions within the tree, resulting in tighter bounding boxes. This paper might be something to explore later (http://www.sci.utah.edu/~wald/Publications/2007/AsyncBuild/asyncbuild.pdf).

Every individual object is also currently being wrapped in a bounding box. I wonder if it's more useful to do this wrapping as an exception rather than as a rule. If a sphere is bound by a tight cube (ie, each of the 6 walls are tangent to the sphere at some point), then a decent amount of times this double hit check on the tight cube and then the sphere is excessive. Maybe this runtime double hit cost could be avoided in favor of just checking the sphere itself. Currently, spheres are wrapped in cubed 100% of the time, but what if it was only done when a given sphere was just really far from the other spheres (ie, total bound volume within boudning volume being used by hittables < some percentage as an example). This would mean skipping 1 hit calculation on objects that are near other objects (ie, the ones that will also more likely have more complex ray interactions too, i think).
Maybe an idea for later.